+++ script/ubuntu-kaiten-yaki.sh
@@ -75,6 +75,13 @@ function main() {
 	PASSPHRASE=""
 	export PASSPHRASE
 
+	# Finishing message
+	cat <<- HEREDOC
+	****************** Install process finished ******************
+
+	...Ready to reboot.
+	HEREDOC
+
 	# Normal end
 	return 0